#28 move waitlist view to its own app
